Use legacy URL to get back to old invoice format
Despite having had the old text format since something like 2012,
earlier this year Amazon decided it was high time to replace it with a
fun web-y version that's much harder to process and frankly just less
useful overall. Great decision.

After months of it being broken, I was just about to rewrite this
scraper when I saw this:
https://www.reddit.com/r/amazonprime/comments/1ly0q5n/new_2025_view_invoice_is_now_graphical_instead_of/
Thanks Reddit!